tests/db_utils: cast numeric columns to float in round_for_comparison
write_outputs.py uses float_format='%.8g' so integer-valued floats are
emitted to CSV as '5' (without trailing '.0').  pd.read_csv then
infers int64 on a column that v3.32.0's golden has as float64 (the
older writer didn't strip the decimal).  Without an explicit cast in
the test's round_for_comparison, the dtype mismatch fails
assert_frame_equal even when the values match — first surfaced by
5weeks_battery_intraperiod_blocks/node_state__dt.csv.

Cast numeric columns to float64 inside round_for_comparison so the
two sides converge to a common dtype regardless of which writer
emitted which '.0'.  Pure test-side widening; matches the spirit of
the comparison gate (tolerate solver-noise + writer-format quirks,
not real value drift).

+++ tests/db_utils.py
@@ -96,7 +96,13 @@ def round_for_comparison(df: pd.DataFrame) -> pd.DataFrame:
     """
     numeric_cols = df.select_dtypes(include="number").columns
     df = df.copy()
-    df[numeric_cols] = df[numeric_cols].round(4)
+    # Cast numeric columns to float64.  ``write_outputs`` uses
+    # ``float_format='%.8g'`` so integer-valued floats are emitted as
+    # ``5`` (no trailing ``.0``); pd.read_csv then infers int64.  The
+    # v3.32.0 goldens were emitted through a path that left ``5.0`` in
+    # place (float64).  Without this cast a column whose all values are
+    # whole numbers fails dtype-equality even when the values match.
+    df[numeric_cols] = df[numeric_cols].astype(float).round(4)
     return df
